requirements.txt
flask
flask-cors
transformers
torch
scikit-learn
youtube-transcript-api
supadata
requests
yt-dlp
webvtt-py
faster-whisper